refactor: cria pasta utils e adiciona i18next
