+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom'
 import './index.css'
 import App from './App'
 // import reportWebVitals from './reportWebVitals'
-import './i18next/index'
+import './utils/i18next'
 
 ReactDOM.render(
   <React.StrictMode>
